внесены доработки в задание
